+++ src-tauri/src/cmd/runtime.rs
@@ -1,0 +1,39 @@
+use crate::{
+    config::*,
+    wrap_err,
+};
+use super::CmdResult;
+use anyhow::Context;
+use std::collections::HashMap;
+use serde_yaml::Mapping;
+
+/// 获取运行时配置
+#[tauri::command]
+pub fn get_runtime_config() -> CmdResult<Option<Mapping>> {
+    Ok(Config::runtime().latest().config.clone())
+}
+
+/// 获取运行时YAML配置
+#[tauri::command]
+pub fn get_runtime_yaml() -> CmdResult<String> {
+    let runtime = Config::runtime();
+    let runtime = runtime.latest();
+    let config = runtime.config.as_ref();
+    wrap_err!(config
+        .ok_or(anyhow::anyhow!("failed to parse config to yaml file"))
+        .and_then(
+            |config| serde_yaml::to_string(config).context("failed to convert config to yaml")
+        ))
+}
+
+/// 获取运行时存在的键
+#[tauri::command]
+pub fn get_runtime_exists() -> CmdResult<Vec<String>> {
+    Ok(Config::runtime().latest().exists_keys.clone())
+}
+
+/// 获取运行时日志
+#[tauri::command]
+pub fn get_runtime_logs() -> CmdResult<HashMap<String, Vec<(String, String)>>> {
+    Ok(Config::runtime().latest().chain_logs.clone())
+}
