+++ src-tauri/src/lib.rs
@@ -1,4 +1,4 @@
-mod cmds;
+mod cmd;
 mod config;
 mod core;
 mod enhance;
@@ -9,7 +9,7 @@ use crate::utils::{resolve, resolve::resolve_scheme, server};
 use config::Config;
 use tauri_plugin_autostart::MacosLauncher;
 use tauri_plugin_deep_link::DeepLinkExt;
-use std::{sync::{Mutex, Once}};
+use std::sync::{Mutex, Once};
 use tauri::AppHandle;
 
 /// A global singleton handle to the application.
@@ -134,61 +134,61 @@ pub fn run() {
         })
         .invoke_handler(tauri::generate_handler![
             // common
-            cmds::get_sys_proxy,
-            cmds::get_auto_proxy,
-            cmds::open_app_dir,
-            cmds::open_logs_dir,
-            cmds::open_web_url,
-            cmds::open_core_dir,
-            cmds::get_portable_flag,
-            cmds::get_network_interfaces,
-            cmds::restart_core,
-            cmds::restart_app,
+            cmd::get_sys_proxy,
+            cmd::get_auto_proxy,
+            cmd::open_app_dir,
+            cmd::open_logs_dir,
+            cmd::open_web_url,
+            cmd::open_core_dir,
+            cmd::get_portable_flag,
+            cmd::get_network_interfaces,
+            cmd::restart_core,
+            cmd::restart_app,
             // clash
-            cmds::get_clash_info,
-            cmds::patch_clash_config,
-            cmds::patch_clash_mode,
-            cmds::change_clash_core,
-            cmds::get_runtime_config,
-            cmds::get_runtime_yaml,
-            cmds::get_runtime_exists,
-            cmds::get_runtime_logs,
-            cmds::uwp::invoke_uwp_tool,
-            cmds::copy_clash_env,
+            cmd::get_clash_info,
+            cmd::patch_clash_config,
+            cmd::patch_clash_mode,
+            cmd::change_clash_core,
+            cmd::get_runtime_config,
+            cmd::get_runtime_yaml,
+            cmd::get_runtime_exists,
+            cmd::get_runtime_logs,
+            cmd::invoke_uwp_tool,
+            cmd::copy_clash_env,
             // verge
-            cmds::get_verge_config,
-            cmds::patch_verge_config,
-            cmds::test_delay,
-            cmds::get_app_dir,
-            cmds::copy_icon_file,
-            cmds::download_icon_cache,
-            cmds::open_devtools,
-            cmds::exit_app,
-            cmds::get_network_interfaces_info,
+            cmd::get_verge_config,
+            cmd::patch_verge_config,
+            cmd::test_delay,
+            cmd::get_app_dir,
+            cmd::copy_icon_file,
+            cmd::download_icon_cache,
+            cmd::open_devtools,
+            cmd::exit_app,
+            cmd::get_network_interfaces_info,
             // profile
-            cmds::get_profiles,
-            cmds::enhance_profiles,
-            cmds::patch_profiles_config,
-            cmds::view_profile,
-            cmds::patch_profile,
-            cmds::create_profile,
-            cmds::import_profile,
-            cmds::reorder_profile,
-            cmds::update_profile,
-            cmds::delete_profile,
-            cmds::read_profile_file,
-            cmds::save_profile_file,
+            cmd::get_profiles,
+            cmd::enhance_profiles,
+            cmd::patch_profiles_config,
+            cmd::view_profile,
+            cmd::patch_profile,
+            cmd::create_profile,
+            cmd::import_profile,
+            cmd::reorder_profile,
+            cmd::update_profile,
+            cmd::delete_profile,
+            cmd::read_profile_file,
+            cmd::save_profile_file,
             // script validation
-            cmds::script_validate_notice,
-            cmds::validate_script_file,
+            cmd::script_validate_notice,
+            cmd::validate_script_file,
             // clash api
-            cmds::clash_api_get_proxy_delay,
+            cmd::clash_api_get_proxy_delay,
             // backup
-            cmds::create_webdav_backup,
-            cmds::save_webdav_config,
-            cmds::list_webdav_backup,
-            cmds::delete_webdav_backup,
-            cmds::restore_webdav_backup,
+            cmd::create_webdav_backup,
+            cmd::save_webdav_config,
+            cmd::list_webdav_backup,
+            cmd::delete_webdav_backup,
+            cmd::restore_webdav_backup,
         ]);
 
     #[cfg(debug_assertions)]
